myapp/requirements.txt
pyautogen
openai
flaml[automl]
